requirements.txt
feedparser==6.0.11
httpx==0.27.0
flask==3.0.3
playwright==1.44.0
python-dotenv==1.0.1
APScheduler==3.10.4
gunicorn==22.0.0
Pillow==10.3.0
ddgs
